Modularize Build
