+++ Makefile
@@ -13,16 +13,17 @@ src/%.o: src/%.c
 bin/s3bd: src/s3bd.o src/cmdline.o
 	$(CC) $(LDFLAGS) $^ -ldl `pkg-config fuse --cflags --libs` -o $@
 
-src/local/%.o: src/local/%.c src/local/%.h
-	$(CC) $(CFLAGS) $< -fpic -fPIC `pkg-config fuse --cflags` -c -o $@
-
-lib/libs3bd_local.so: src/local/callbacks.o
-	$(CC) $(CFLAGS) $< -shared -o $@
+lib/libs3bd_local.so:
+	CC=$(CC) CFLAGS="$(CFLAGS)" make -C src/backends/local
+	mv -f src/backends/local/libs3bd_local.so lib/libs3bd_local.so
 
 clean:
-	rm -f src/*.o src/local/*.o
+	rm -f src/*.o
+	make -C src/backends/local clean
 
 cleaner: clean
 	rm -f bin/s3bd lib/*.so
+	make -C src/backends/local cleaner
 
 cleanest: cleaner
+	make -C src/backends/local cleanest
